+++ src/admin-supply-line-list/admin-supply-line-list.module.js
@@ -28,6 +28,7 @@
         'openlmis-pagination',
         'openlmis-rights',
         'referencedata-facility',
+        'referencedata-feature-flag',
         'referencedata-supply-line',
         'referencedata-requisition-group',
         'openlmis-object-utils',